server/requirements.txt
aiohappyeyeballs==2.6.1
anthropic==0.76.0
aiohttp==3.13.3
aiohttp-retry==2.9.1
aiosignal==1.4.0
attrs==25.4.0
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
Flask==3.1.2
frozenlist==1.8.0
google-api-core==2.29.0
google-auth==2.47.0
google-cloud-core==2.5.0
google-cloud-datastore==2.23.0
google-cloud-ndb==2.4.0
google-cloud-storage==2.19.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
Pillow==11.1.0
propcache==0.4.1
proto-plus==1.27.0
protobuf==6.33.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
pymemcache==4.0.0
pytz==2025.2
redis==6.4.0
requests==2.32.5
rsa==4.9.1
twilio==9.9.1
typing_extensions==4.15.0
urllib3==2.6.3
Werkzeug==3.1.5
yarl==1.22.0
pytest==8.3.5
